# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cell Messenger API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Cell API Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Прямое подключение к Go серверу (локально или контейнер API)"
    },
    {
      "url": "http://localhost",
      "description": "Подключение через Nginx Reverse Proxy"
    },
    {
      "url": "/",
      "description": "Относительный корень сервера"
    }
  ],
...
  "paths": {
...
    "/api/v1/ws/": {
      "get": {
        "tags": [
          "WebSocket"
        ],
        "summary": "Подключение к WebSocket серверу",
        "description": "Инициализирует WebSocket-соединение для обмена сообщениями, индикации набора текста и подтверждений в реальном времени.\n\n### Аутентификация\nJWT токен передается в строке запроса в параметре `token`:\n`GET /api/v1/ws/?token=<JWT_TOKEN>`\n\n---\n\n### Протокол взаимодействия\n\nВсе пакеты передаются в формате JSON.\n\n#### 1. Клиент -> Сервер (Client Actions)\nКорневой конверт:\n```json\n{\n  \"action\": \"<string>\",\n  \"payload\": { ... }\n}\n```\n\n* **`message:send`** — отправка сообщения в чат:\n```json\n{\n  \"action\": \"message:send\",\n  \"payload\": {\n    \"client_msg_temp_id\": \"temp_1712000000\",\n    \"chat_id\": \"a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d\",\n    \"content\": \"Привет, как дела?\"\n  }\n}\n```\n\n* **`typing`** — отправка индикации набора текста:\n```json\n{\n  \"action\": \"typing\",\n  \"payload\": {\n    \"user_id\": \"550e8400-e29b-41d4-a716-446655440000\",\n    \"chat_id\": \"a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d\",\n    \"is_typing\": true\n  }\n}\n```\n\n* **`read`** — отметка о прочтении сообщений до указанного ID:\n```json\n{\n  \"action\": \"read\",\n  \"payload\": {\n    \"chat_id\": \"a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d\",\n    \"message_id\": 142\n  }\n}\n```\n\n---\n\n#### 2. Сервер -> Клиент (Server Events)\nКорневой конверт:\n```json\n{\n  \"event\": \"<string>\",\n  \"payload\": { ... }\n}\n```\n\n* **`message:ack`** — подтверждение автору о сохранении сообщения в БД:\n```json\n{\n  \"event\": \"message:ack\",\n  \"payload\": {\n    \"client_msg_temp_id\": \"temp_1712000000\",\n    \"message_id\": 204,\n    \"chat_id\": \"a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d\",\n    \"created_at\": \"2026-10-03T15:00:00Z\"\n  }\n}\n```\n\n* **`message:new`** — широковещательное уведомление собеседникам о новом сообщении:\n```json\n{\n  \"event\": \"message:new\",\n  \"payload\": {\n    \"id\": 204,\n    \"chat_id\": \"a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d\",\n    \"sender_id\": \"550e8400-e29b-41d4-a716-446655440000\",\n    \"content\": \"Привет, как дела?\",\n    \"created_at\": \"2026-10-03T15:00:00Z\"\n  }\n}\n```\n\n* **`typing`** — трансляция статуса набора текста участникам чата:\n```json\n{\n  \"event\": \"typing\",\n  \"payload\": {\n    \"user_id\": \"550e8400-e29b-41d4-a716-446655440000\",\n    \"chat_id\": \"a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d\",\n    \"is_typing\": true\n  }\n}\n```\n\n* **`read`** — трансляция обновления курсора прочтения:\n```json\n{\n  \"event\": \"read\",\n  \"payload\": {\n    \"user_id\": \"550e8400-e29b-41d4-a716-446655440000\",\n    \"chat_id\": \"a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d\",\n    \"message_id\": 142\n  }\n}\n```\n\n* **`error`** — ошибка обработки действия:\n```json\n{\n  \"event\": \"error\",\n  \"payload\": {\n    \"client_msg_temp_id\": \"temp_1712000000\",\n    \"error\": \"forbidden: you are not a member of this chat\"\n  }\n}\n```\n",
        "operationId": "connectWebSocket",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "description": "JWT токен авторизации пользователя",
            "schema": {
              "type": "string"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        ],
        "responses": {
          "101": {
            "description": "Протокол успешно переключен (Switching Protocols to WebSocket)"
          },
          "400": {
            "description": "Ошибка WebSocket апгрейда",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "failed to upgrade to websocket"
                }
              }
            }
          },
          "401": {
            "description": "Токен не предоставлен или не прошел валидацию",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingToken": {
                    "summary": "Токен не передан",
                    "value": {
                      "error": "access token not found"
                    }
                  },
                  "invalidToken": {
                    "summary": "Недействительный токен",
                    "value": {
                      "error": "invalid token"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calls/rooms": {
      "post": {
        "tags": [
          "Calls"
        ],
        "summary": "Создание или получение активной комнаты звонка для чата",
        "description": "Создает новую комнату для аудио/видео звонка в указанном чате или возвращает уже существующую активную комнату. Доступно только участникам чата.",
        "operationId": "createCallRoom",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCallRoomRequest"
              },
              "example": {
                "chat_id": "550e8400-e29b-41d4-a716-446655440000",
                "type": "video"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Комната успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallRoomResponse"
                }
              }
            }
          },
          "200": {
            "description": "Комната уже существует и активна в данном чате",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallRoomResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос или отсутствие chat_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не является участником данного чата",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/calls/active": {
      "get": {
        "tags": [
          "Calls"
        ],
        "summary": "Проверка активного звонка в чате",
        "description": "Проверяет, идет ли в данный момент звонок в указанном чате, и возвращает информацию о комнате и количестве подключенных участников.",
        "operationId": "getActiveCall",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "query",
            "required": true,
            "description": "UUID чата для проверки",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Статус звонка в чате",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveCallResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Пользователь не состоит в чате",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/calls/ws": {
      "get": {
        "tags": [
          "Calls"
        ],
        "summary": "Сигнальный WebSocket для WebRTC SFU звонков",
        "description": "Устанавливает полнодуплексное WebSocket соединение для сигнального протокола WebRTC. Обеспечивает обмен SDP Offer/Answer, ICE кандидатами и событиями входа/выхода участников.\n\n### Проверка авторизации\n* Передается `token` в query-параметре или заголовок `Authorization: Bearer <token>`.\n* Сервис проверяет членство пользователя в указанном чате (`chat_id`) через PostgreSQL.\n* Если пользователь не состоит в чате — соединение отклоняется с кодом 403 Forbidden.",
        "operationId": "connectCallsWebSocket",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "description": "JWT токен авторизации",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chat_id",
            "in": "query",
            "required": false,
            "description": "UUID чата (обязателен, если не указан room_id)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "query",
            "required": false,
            "description": "UUID комнаты звонка",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "description": "Тип звонка при автоматическом создании ('video' или 'audio')",
            "schema": {
              "type": "string",
              "enum": [
                "video",
                "audio"
              ],
              "default": "video"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Протокол успешно переключен на WebSocket (WebRTC SFU)"
          },
          "400": {
            "description": "Не указаны chat_id или room_id"
          },
          "401": {
            "description": "Токен не предоставлен или недействителен"
          },
          "403": {
            "description": "Пользователь не является участником указанного чата"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Авторизационный JWT токен. Передается в заголовке `Authorization: Bearer <token>`"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Описание ошибки",
            "example": "invalid request payload"
          },
          "sys": {
            "type": "string",
            "description": "Системное/детализированное сообщение об ошибке (при наличии)",
            "example": "context deadline exceeded"
          }
        }
      },
...
      "CreateCallRoomRequest": {
        "type": "object",
        "required": [
          "chat_id"
        ],
        "properties": {
          "chat_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID чата, в котором инициируется звонок",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "type": {
            "type": "string",
            "enum": [
              "video",
              "audio"
            ],
            "default": "video",
            "description": "Тип звонка: видео или аудио",
            "example": "video"
          }
        }
      },
      "CallParticipant": {
        "type": "object",
        "required": [
          "peer_id",
          "user_id",
          "username",
          "joined_at"
        ],
        "properties": {
          "peer_id": {
            "type": "string",
            "description": "Уникальный идентификатор пира в сессии SFU",
            "example": "a8f3b2c1d4e5"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID пользователя",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "username": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "alex_dev"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time",
            "description": "Время подключения к комнате",
            "example": "2026-10-05T07:30:00Z"
          }
        }
      },
      "CallRoomResponse": {
        "type": "object",
        "required": [
          "room_id",
          "chat_id",
          "call_type",
          "initiator_id",
          "created_at",
          "is_active",
          "participants"
        ],
        "properties": {
          "room_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID комнаты звонка",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "chat_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID чата",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "call_type": {
            "type": "string",
            "enum": [
              "video",
              "audio"
            ],
            "description": "Тип звонка",
            "example": "video"
          },
          "initiator_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID пользователя-инициатора",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "initiator_name": {
            "type": "string",
            "description": "Имя инициатора звонка",
            "example": "alex_dev"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Время создания комнаты",
            "example": "2026-10-05T07:30:00Z"
          },
          "is_active": {
            "type": "boolean",
            "description": "Флаг активности комнаты звонка",
            "example": true
          },
          "participants": {
            "type": "array",
            "description": "Список текущих участников звонка",
            "items": {
              "$ref": "#/components/schemas/CallParticipant"
            }
          }
        }
      },
      "ActiveCallResponse": {
        "type": "object",
        "required": [
          "active"
        ],
        "properties": {
          "active": {
            "type": "boolean",
            "description": "Признак наличия активного звонка в чате",
            "example": true
          },
          "chat_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID чата",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "room": {
            "$ref": "#/components/schemas/CallRoomResponse"
          }
        }
      },
...
    }
  }
}